+++ src/logic/constants.tsx
@@ -1,0 +1,25 @@
+// TODO voir comment gérer les différentes législatures
+export const CURRENT_LEGISLATURE = 16
+
+// Colors copied from conf/app.yml in nosdeputes.fr
+export const GroupeColorsByAcronyme: { [k: string]: string } = {
+  LFI: rgbToHex(204, 42, 70),
+  GDR: rgbToHex(207, 77, 39),
+  SOC: rgbToHex(255, 149, 145),
+  ECO: rgbToHex(151, 215, 74),
+  LIOT: rgbToHex(216, 226, 24),
+  REN: rgbToHex(255, 200, 0),
+  MODEM: rgbToHex(255, 152, 0),
+  HOR: rgbToHex(55, 157, 200),
+  LR: rgbToHex(78, 81, 212),
+  RN: rgbToHex(19, 57, 62),
+  NI: rgbToHex(165, 165, 165),
+}
+
+function rgbToHex(r: number, g: number, b: number) {
+  function componentToHex(c: number) {
+    var hex = c.toString(16)
+    return hex.length == 1 ? '0' + hex : hex
+  }
+  return '#' + componentToHex(r) + componentToHex(g) + componentToHex(b)
+}
